+++ app/components/YourBoards.vue
@@ -7,6 +7,7 @@
             :id="item.id"
             :name="item.name"
             :style="item.style"
+            :image="item.image"
         />
         <button
             type="button"